Task1. Created a string, using String, consisting of numbers from 0 to 20000.
Created a string, using StringBuilder, consisting of numbers from 0 to 20000.

In order to feel the difference in performance between the concatenation of ordinary strings (String class) and the use of StringBuilder, I implemented the task in these two ways, measuring the program's running time.

Task 2.
Copied the Aircraft class from the 6 day challenge.
Made changes to the class so that the following code displays information about the aircraft, similar to calling the info () method.