1. Created the Human class. Added the name field to Human. I added a constructor, get and set methods to this field in the class, I also created the printInfo () method, which prints information to the console about Human in the format: “This person is named NAME”.
2. Created a class Student, which is inherited from the class
human. Student added an additional string field - nameGroup. For this field, I also created get and set methods. The constructor in the Student class must take two arguments as input - name and nameGroup. The printInfo() method has been overridden so that the format of the message output to the console is as follows:
“This person named NAME”
“This student named NAME”
Used the super keyword.

3. Created the PhysicsTeacher class. PhysicsTeacher also inherits from the Human class. PhysicsTeacher has added an additional string field - nameLesson, which is maintained by this teacher. Created get and set methods for this field. The constructor in the PhysicsTeacher class takes two inputs
argument - the name of the teacher and the name of the subject being taught. The printInfo() method has been overridden so that the format of the message output to the console is as follows:
“This person named NAME”
“This teacher is named NAME.”
Used the super keyword.

4.Created an object of the Student class and an object of the class in the main() method of the Task1 class
physics teacher. Displayed in the console the nameGroup of the student object and
nameLesson on the teacher object. Then, called printInfo() on
objects.