Implemented a class - TestFigures. In this class, in the main() method, I created and placed in an array the following geometric figures: Figure[] figures = { new Triangle(10, 10, 10, "Red"), new Triangle(10, 20, 30, "Green"), new Triangle(10, 20, 15, "Red"), new Rectangle(5, 10, "Red"), new Rectangle(40, 15, "orange"), new Circle(4, "Red"), new Circle(10, "Red"), new Circle(5, "Blue") };
In the same class, I implemented two methods:
public static double calculateRedPerimeter(Figure[] figures)
public static double calculateRedArea(Figure[] figures)

The first method, taking an array of geometric shapes as input, returns
the sum of the perimeters of the red figures. The second method, taking an array as input
geometric shapes, returns the sum of the areas of the red shapes.
Called these two methods on the figures array and printed the result to the console.