1. Created a class: Warehouse with fields - countOrder, balance. Get and set methods for both fields. To get information about the values of the warehouse fields, I overridden the toString() method.
2.Created a class: Picker, with salary field, only get method.
Courier, with salary field, only get method.

3. In each employee class, I implemented the Worker interface, in which I declared the doWork() and bonus() methods.
Each time an employee performs his work (calling the doWork() method), he is paid wages (80 - Picker, 100 - Courier).
Also, when calling doWork() on the Picker, the value of the countOrder field in the Warehouse class is incremented by 1.
When calling doWork() on Courier, the balance variable in the Warehouse class is incremented by 1000.
Employees are entitled to a bonus (individual calculation for each): when in stock
The 1500th order is shipped, and the Picker's salary up to that point is tripled. When
the warehouse earned 1.000.000, the Courier's salary earned by that time is doubled.

4. To demonstrate and test the operation of the program, I created a warehouse and 1 worker each,
called the doWork() and bonus() methods on each, so many times that each of
employees received a bonus.
Displayed the balance and the number of completed orders in the warehouse and the salary of each employee in the console.
To get warehouse data inside the employee classes, I implemented storing a reference to the Warehouse object and passing it using the constructor.