1. Created a chalkboard. Implemented a two-dimensional array ShipArray. Implemented a constructor, getting and setting methods for it. Also implemented a print method for output to the console.
2.Created an enumeration of Ships.
Implemented enumeration of empty space symbols, halos around ships,
sos signal, ship from 4-deck to 1-deck, as well as misses and hits on the game board using symbols. Added string and integer fields, implemented for them by the constructor, get and toString methods.

3. Created the ShipCommander class.
Added board1, board2 and name fields. Implemented a constructor for the class, getting and printing methods.
A map method has been created to display the exit board to the console.
Created an integer array moveCoord. I implemented it using a Scanner to enter shot coordinates onto the board. Added exceptions: the number was from 0 to 9; So that you cannot enter the same coordinates repeatedly; So that the entered number is an integer. Implemented turn transition on miss.
A method of movement has been created, in it with the help of, if all the outcomes of the shots have been realized (Wounded, Killed, Missed).
Using the random implementation method, the first move was determined. The winner method uses a loop to determine the winner if certain standard conditions are met.

4. The Main class implements the main method.
In the ShipArrayEmpty method I create a two-dimensional array Ships[10][10]. Using loops, I go through a two-dimensional array and check which cells are empty (EMPTY).
Next, I create methods for all ships, with the help of which I write the coordinates of ships into an array, starting from 4-deck to 1-deck. I'm adding exceptions for entering coordinates using the sights of a horizontal or vertical ship.
In the addOreol method, using loops for and conditions, if I create a halo in one cell in each ship, since ships cannot be placed close to each other.
The main method implements all the above methods.